Implemented a new method to identify vendor:
At first, we don't need to have a correct vendor since the software name must match anyway during CPE search.

As the CPE dictionary as a (SOFT,VENDOR) table, if we guess the software name correctly we can just look its
vendors (there may be severals depending on versions) up with a simple "SELECT VENDOR WHERE SOFT='soft_name'".

Only if the result of this SELECT returns several vendors do we need to use the provided vendor name to try to
identify the correct one.

The first part of this algorithm is implemented by this commit.

The second part (finding the correct vendor in case the SELECT returns several vendors) must still be implemented.